don't test matplotlib installed __init__ file at configure time (we don't know where it will be by this time and it is not there yet anyway). Test in the python patch script instread - this is just a partial solution.
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/development/brainvisa-cmake/branches/bug_fix@82123 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ cmake/brainvisa-packaging-python-matplotlib-patchinit.py
@@ -5,15 +5,20 @@ import os, re, sys
 infile = sys.argv[1]
 outfile = sys.argv[2]
 
-r = re.compile( "((^.*$\n)*)^( *path = '/usr/share/matplotlib/mpl-data')$\n((^.*$\n)*)^( *raise RuntimeError\('Could not find the matplotlib data files'\))$\n+( *# setuptools' namespace_packages may highjack this init file$\n(^.*$\n)*)^( *path = *('/etc') +# guaranteed to exist or raise)$\n((^.*$\n?)*)\Z", re.MULTILINE )
+if not os.path.exists(infile):
+    # matplotlib __init__ file is not here, for now we do nothing (probably
+    # needs a better fix one day)
+    sys.exit(0)
 
-lines = open( infile ).read()
-m = r.match( lines )
+r = re.compile("((^.*$\n)*)^( *path = '/usr/share/matplotlib/mpl-data')$\n((^.*$\n)*)^( *raise RuntimeError\('Could not find the matplotlib data files'\))$\n+( *# setuptools' namespace_packages may highjack this init file$\n(^.*$\n)*)^( *path = *('/etc') +# guaranteed to exist or raise)$\n((^.*$\n?)*)\Z", re.MULTILINE)
 
-olines = open( outfile, 'w' )
-olines.write( m.group(1) )
-olines.write( "    path = os.sep.join([os.path.dirname(__file__), 'mpl-data'])\n" )
-olines.write( m.group(4) )
-olines.write( m.group(7) )
-olines.write( '    path =  get_data_path() # guaranteed to exist or raise\n' )
-olines.write( m.group(11) )
+lines = open(infile).read()
+m = r.match(lines)
+
+olines = open( outfile, 'w')
+olines.write(m.group(1))
+olines.write( "    path = os.sep.join([os.path.dirname(__file__), 'mpl-data'])\n")
+olines.write(m.group(4))
+olines.write(m.group(7))
+olines.write('    path =  get_data_path() # guaranteed to exist or raise\n')
+olines.write(m.group(11))
